Fixed per test coverage
