+++ test/evergreen.yml
@@ -3012,13 +3012,15 @@ tasks:
 
             # Record the end time, in seconds
             date +%s >> time.txt
+
+            rm -rf build_*
       - func: "code coverage analysis"
         vars:
           generate_atlas_format: true
           working_dir: "wiredtiger"
       - func: "upload artifact"
         vars:
-          upload_source_dir: "wiredtiger/coverage_report"
+          upload_source_dir: "wiredtiger"
           upload_filename: coverage-report-python.tgz
       - func: "cleanup"
       
@@ -3046,13 +3048,15 @@ tasks:
 
             # Record the end time, in seconds
             date +%s >> time.txt
+
+            rm -rf build_*
       - func: "code coverage analysis"
         vars:
           generate_atlas_format: true
           working_dir: "wiredtiger"
       - func: "upload artifact"
         vars:
-          upload_source_dir: "wiredtiger/coverage_report"
+          upload_source_dir: "wiredtiger"
           upload_filename: coverage-report-csuite.tgz
       - func: "cleanup"
   
@@ -3073,7 +3077,6 @@ tasks:
           dependent_task: coverage-report-csuite
           destination: coverage_report_csuite
       - command: shell.exec
-        working_dir: wiredtiger
         params:
           shell: bash
           script: |
@@ -3085,8 +3088,8 @@ tasks:
             source venv/bin/activate
             pip3 install lxml==4.8.0 Pygments==2.11.2 Jinja2==3.0.3 gcovr==5.0
             mkdir -p full_coverage_report
-            GCOV=/opt//mongodbtoolchain/v4/bin/gcov gcovr --add-tracefile coverage_report_csuite/full_coverage_report.json --add-tracefile coverage_report_python/full_coverage_report.json -j 4 --html-self-contained --html-details full_coverage_report/2_coverage_report.html --json-summary-pretty --json-summary full_coverage_report/1_coverage_report_summary.json
-            ${python_binary|python3} test/evergreen/code_coverage_analysis.py -s full_coverage_report/1_coverage_report_summary.json -t time.txt
+            gcovr -r coverage_report_csuite --add-tracefile coverage_report_csuite/coverage_report/full_coverage_report.json --add-tracefile coverage_report_python/coverage_report/full_coverage_report.json -j 4 --html-self-contained --html-details full_coverage_report/2_coverage_report.html --json-summary-pretty --json-summary full_coverage_report/1_coverage_report_summary.json
+            ${python_binary|python3} coverage_report_csuite/test/evergreen/code_coverage_analysis.py -s full_coverage_report/1_coverage_report_summary.json -t time.txt
       - func: "upload stats to atlas"
         vars:
           stats_dir: ./full_coverage_report
